refactor: added components to seperate folder
