+++ src/utils/labels.js
@@ -7,7 +7,8 @@ const labels = {
     ADD_TO_WISHLIST: "ADD_TO_WISHLIST",
     INCREMENT_QUANTITY: "INCREMNT_QUANTITY",
     DECREMENT_QUANTITY: "DECREMENT_QUANTITY",
-    REMOVE_FROM_WISHLIST: "REMOVE_FROM_WISHLIST"
+    REMOVE_FROM_WISHLIST: "REMOVE_FROM_WISHLIST",
+    REMOVE_ITEM: "REMOVE_ITEM"
 }
 
 export { labels };
